requirements.txt
# Django Core
Django==5.0.3
django-ninja==1.2.2
django-ninja-jwt==5.3.3
django-ninja-extra==0.21.2

# Database
psycopg2-binary==2.9.9

# Cache & Queue
redis==5.0.3
celery==5.3.6
django-redis==5.4.0

# Real-time
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Image Processing
Pillow==10.2.0

# Markdown & Content
markdown==3.5.2
bleach==6.1.0
pymdown-extensions==10.7

# Utilities
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3

# Validation
pydantic==2.6.3

# Date & Time
python-dateutil==2.9.0

# HTTP Client
httpx==0.27.0

# Development
django-debug-toolbar==4.3.0
ipython==8.22.1

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==24.2.0
flake8==7.0.0
isort==5.13.2
